requirements.txt
cffi==1.16.0
cryptography==41.0.7
dnspython==2.4.2
pycparser==2.21
validate-email-address==1
